+++ flow-typed/baum.js
@@ -18,6 +18,8 @@ declare module '@prostory/baum' {
     toNotEqual: (expected: mixed) => void,
     toThrow: (expectedError?: Error) => void,
     toNotThrow: () => void,
+    toMatch: (expected: string | RegExp) => void,
+    toNotMatch: (expected: string | RegExp) => void,
     toBeResolved: () => Promise<ExpectChecks>,
     toBeRejected: (expectedError?: Error) => Promise<void>,
   }